+++ app/android/src/main/java/club/nito/app/NitoApp.kt
@@ -6,8 +6,15 @@ import androidx.compose.material3.Surface
 import androidx.compose.material3.windowsizeclass.WindowSizeClass
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import club.nito.app.di.nitoDateTimeFormatterModule
+import club.nito.app.di.userMessageStateHolderModule
+import club.nito.core.data.di.dataModule
 import club.nito.core.designsystem.theme.NitoTheme
+import club.nito.core.domain.di.useCaseModule
 import club.nito.core.model.AuthStatus
+import club.nito.core.network.di.remoteDataSourceModule
+import club.nito.core.network.di.supabaseClientModule
+import org.koin.compose.KoinApplication
 
 @Composable
 fun NitoApp(
@@ -15,15 +22,33 @@ fun NitoApp(
     authStatus: AuthStatus?,
     modifier: Modifier = Modifier,
 ) {
-    NitoTheme {
-        Surface(
-            modifier = modifier.fillMaxSize(),
-            color = MaterialTheme.colorScheme.background,
-        ) {
-            NitoNavHost(
-                windowSize = windowSize,
-                authStatus = authStatus,
+    KoinApplication(
+        application = {
+            modules(
+                nitoDateTimeFormatterModule,
+                userMessageStateHolderModule,
+
+                supabaseClientModule,
+                remoteDataSourceModule,
+//                fakeRemoteDataSourceModule,
+                dataModule,
+                useCaseModule,
             )
+
+//            val kermit = Logger.withTag("koin")
+//            logger(KermitKoinLogger(kermit))
+        },
+    ) {
+        NitoTheme {
+            Surface(
+                modifier = modifier.fillMaxSize(),
+                color = MaterialTheme.colorScheme.background,
+            ) {
+                NitoNavHost(
+                    windowSize = windowSize,
+                    authStatus = authStatus,
+                )
+            }
         }
     }
 }
