+++ app/android/src/main/java/club/nito/app/di/UserMessageStateHolderModule.kt
@@ -6,10 +6,16 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import org.koin.core.module.dsl.singleOf
+import org.koin.dsl.module
 
 @Module
 @InstallIn(SingletonComponent::class)
 class UserMessageStateHolderModule {
     @Provides
     fun provideUserMessageStateHolder(): UserMessageStateHolder = DefaultUserMessageStateHolder()
+}
+
+val userMessageStateHolderModule = module {
+    singleOf<UserMessageStateHolder>(::DefaultUserMessageStateHolder)
 }
