+++ app/ios-combined/src/commonMain/kotlin/club/nito/ios/combined/KmpEntryPoint.kt
@@ -1,9 +1,9 @@
 package club.nito.ios.combined
 
-import club.nito.core.data.dataModule
-import club.nito.core.domain.useCaseModule
-import club.nito.core.network.remoteDataSourceModule
-import club.nito.core.network.supabaseClientModule
+import club.nito.core.data.di.dataModule
+import club.nito.core.domain.di.useCaseModule
+import club.nito.core.network.di.remoteDataSourceModule
+import club.nito.core.network.di.supabaseClientModule
 import co.touchlab.kermit.Logger
 import co.touchlab.kermit.koin.KermitKoinLogger
 import org.koin.core.KoinApplication
