+++ core/common/src/commonMain/kotlin/club/nito/core/common/NitoDateTimeFormatter.kt
@@ -1,0 +1,10 @@
+package club.nito.core.common
+
+import kotlinx.datetime.Instant
+
+interface NitoDateTimeFormatter {
+    /**
+     * 日時をフォーマットする
+     */
+    fun formatDateTimeString(instant: Instant): String
+}
