+++ core/data/build.gradle.kts
@@ -18,12 +18,9 @@ kotlin {
                 implementation(projects.core.network)
 
                 implementation(libs.kotlinxCoroutinesCore)
-                implementation(libs.kermit)
-            }
-        }
-        iosMain {
-            dependencies {
+
                 implementation(libs.koin)
+                implementation(libs.kermit)
             }
         }
     }
