+++ core/domain/build.gradle.kts
@@ -17,12 +17,9 @@ kotlin {
                 implementation(projects.core.model)
 
                 implementation(libs.kotlinxCoroutinesCore)
-                implementation(libs.kermit)
-            }
-        }
-        iosMain {
-            dependencies {
+
                 implementation(libs.koin)
+                implementation(libs.kermit)
             }
         }
     }
