--- core/domain/src/iosMain/kotlin/club/nito/core/domain/useCaseModule.kt
@@ -1,16 +1,0 @@
-package club.nito.core.domain
-
-import org.koin.core.module.Module
-import org.koin.core.module.dsl.singleOf
-import org.koin.dsl.bind
-import org.koin.dsl.module
-
-val useCaseModule: Module = module {
-    singleOf(::ObserveAuthStatusExecutor) bind ObserveAuthStatusUseCase::class
-    singleOf(::SignInExecutor) bind SignInUseCase::class
-    singleOf(::ModifyPasswordExecutor) bind ModifyPasswordUseCase::class
-    singleOf(::SignOutExecutor) bind SignOutUseCase::class
-    singleOf(::GetRecentScheduleExecutor) bind GetRecentScheduleUseCase::class
-    singleOf(::GetParticipantScheduleListExecutor) bind GetParticipantScheduleListUseCase::class
-    singleOf(::ParticipateExecutor) bind ParticipateUseCase::class
-}
