+++ gradle/libs.versions.toml
@@ -23,6 +23,7 @@ androidxTestExt = "1.1.5"
 androidxTestEspresso = "3.5.1"
 dagger = "2.48.1"
 koin = "3.4.3"
+koinCompose = "1.1.0"
 ktlint = "0.49.1"
 kotlinxSerialization = "1.6.0"
 ktor = "2.3.4"
@@ -72,6 +73,7 @@ daggerHiltAndroidCompiler = { module = "com.google.dagger:hilt-android-compiler"
 androidxLifecycleProcess = { group = "androidx.lifecycle", name = "lifecycle-process", version.ref = "androidxLifecycleProcess" }
 
 koin = { module = "io.insert-koin:koin-core", version.ref = "koin" }
+koinCompose = { module = "io.insert-koin:koin-compose", version.ref = "koinCompose" }
 
 kotlinxCoroutinesCore = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-core", version.ref = "kotlinxCoroutines" }
